Develop

+++ src/pages/dashboard/index.tsx
@@ -1,54 +1,82 @@
 import { useState, useCallback } from 'react';
 import { NextPage } from 'next';
 import { useSelector, useDispatch } from 'react-redux';
-// import { Container, Col, Row } from 'react-grid-system';
+import swal from 'sweetalert';
 
 import Layout from '../../components/layout/Layout';
 import { Modal, Table } from '../../components/UI';
+import { DataType } from '../../types';
+// import { selectUser } from '../../infrastructure/store/users/userSlice';
 
 import {
-  selectUser,
-  updateName,
-} from '../../infrastructure/store/users/userSlice';
-
-import {
+  syncIssue,
   selectIssues,
-  createIssue,
   deleteIssue,
 } from '../../infrastructure/store/data/dataSlice';
 
 const Dashboard: NextPage = () => {
-  const [isModalOpen, setIsModalOpen] = useState(false);
+  const [filterArr, setFilterArr] = useState<DataType[]>([]);
+  const [isModalOpen, setIsModalOpen] = useState({
+    isOpen: false,
+    isEdit: false,
+  });
 
-  const user = useSelector(selectUser);
+  // const user = useSelector(selectUser);
   const issues = useSelector(selectIssues);
   const dispatch = useDispatch();
 
-  const createIssueHandler = useCallback(() => {
-    setIsModalOpen(!isModalOpen);
-    // dispatch(createIssue(mockData));
+  const syncIssueHandler = useCallback(() => {
+    swal({
+      text: 'Are you sure that you want to sync data?',
+      dangerMode: false,
+    }).then((willSync) => {
+      if (willSync) {
+        dispatch(syncIssue());
+      }
+    });
   }, []);
+
+  const createIssueHandler = useCallback(() => {
+    setIsModalOpen({
+      isOpen: true,
+      isEdit: false,
+    });
+  }, []);
+
+  const editIssueHandler = useCallback(
+    (id) => {
+      setFilterArr(issues.filter((issue) => issue.id === id));
+      setIsModalOpen({
+        isOpen: true,
+        isEdit: true,
+      });
+    },
+    [issues]
+  );
 
   const deleteIssueHandler = useCallback((id) => {
-    dispatch(deleteIssue(id));
+    swal({
+      title: 'Are you sure?',
+      text: 'Are you sure that you want to delete this issue?',
+      icon: 'warning',
+      dangerMode: true,
+    }).then((willDelete) => {
+      if (willDelete) {
+        dispatch(deleteIssue(id));
+      }
+    });
   }, []);
 
-  const demo = () => {
-    dispatch(updateName('Hello'));
-  };
-
   return (
-    <Layout title="demo">
-      <Modal isModalOpen={isModalOpen} />
+    <Layout title="dashboard">
+      <Modal updateIssueArr={filterArr} isModalOpen={isModalOpen} />
       <Table
         issues={issues}
+        syncIssueHandler={syncIssueHandler}
         createIssueHandler={createIssueHandler}
+        editIssueHandler={editIssueHandler}
         deleteIssueHandler={deleteIssueHandler}
       />
-      {/* <h1>{user}</h1>
-        <button onClick={demo} className="bg-red-300">
-          demo
-        </button> */}
     </Layout>
   );
 };
